# backend/artifacts/openapi/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/owner/me": {
      "get": {
        "operationId": "profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Идентификатор, телефон и число питомцев."
          },
          "401": {
            "description": "Bearer token отсутствует или невалиден."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Профиль текущего владельца",
        "tags": [
          "Owner profile"
        ]
      }
    },
...
    "/v1/clinic/{clinicId}/locations/{locationId}/appointments": {
      "get": {
        "operationId": "ClinicAppointmentsRegistry_list",
        "parameters": [
          {
            "name": "clinicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "bucket",
            "required": true,
            "in": "query",
            "schema": {
              "enum": [
                "upcoming",
                "history"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stable cursor page from a PostgreSQL-owned snapshot boundary.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "clinicId",
                    "locationId",
                    "serverNow",
                    "items",
                    "nextCursor"
                  ],
                  "properties": {
                    "clinicId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "locationId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "serverNow": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "nextCursor": {
                      "type": "string",
                      "nullable": true
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "appointmentId",
                          "aggregateVersion",
                          "statusCode",
                          "statusLabel",
                          "slot",
                          "pet",
                          "service"
                        ],
                        "properties": {
                          "appointmentId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "aggregateVersion": {
                            "type": "integer"
                          },
                          "statusCode": {
                            "type": "string",
                            "enum": [
                              "SCHEDULED",
                              "COMPLETED",
                              "NO_SHOW",
                              "CANCELLED"
                            ]
                          },
                          "statusLabel": {
                            "type": "string"
                          },
                          "slot": {
                            "type": "object",
                            "required": [
                              "startsAt",
                              "endsAt"
                            ],
                            "properties": {
                              "startsAt": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "endsAt": {
                                "type": "string",
                                "format": "date-time"
                              }
                            }
                          },
                          "pet": {
                            "type": "object",
                            "required": [
                              "id",
                              "name",
                              "speciesLabel"
                            ],
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "name": {
                                "type": "string"
                              },
                              "speciesLabel": {
                                "type": "string"
                              }
                            }
                          },
                          "service": {
                            "type": "object",
                            "nullable": true,
                            "required": [
                              "displayName"
                            ],
                            "properties": {
                              "displayName": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bucket, limit or cursor is invalid."
          },
          "401": {
            "description": "Clinic employee JWT is required."
          },
          "403": {
            "description": "Capability, clinic, location and active membership are required."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Location-scoped clinic appointments registry",
        "tags": [
          "Clinic Portal"
        ]
      }
    },
...
    "/v1/clinic/{clinicId}/locations/{locationId}/patients/{patientId}": {
      "get": {
        "description": "Default-off rollout. Returns no clinical, owner-contact, document or financial fields.",
        "operationId": "ClinicPatientDetail_detail",
        "parameters": [
          {
            "name": "clinicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicPatientDetailDto"
                }
              }
            }
          },
          "401": {
            "description": "Clinic employee JWT is required."
          },
          "403": {
            "description": "Administrative capability and exact active membership are required."
          },
          "404": {
            "description": "Rollout disabled or patient is not currently visible in the exact scope."
          },
          "500": {
            "description": "Controlled technical or malformed-source failure."
          },
          "503": {
            "description": "Visibility policy configuration is unavailable."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Current administrative patient detail for an exact clinic location",
        "tags": [
          "Clinic Patients"
        ],
        "x-required-capabilities": [
          "patient.admin.read"
        ]
      }
    },
...
    "/v1/clinic/{clinicId}/locations/{locationId}/patients/{patientId}/local-profile/reference": {
      "patch": {
        "operationId": "ClinicPatientAdministrativeReference_update",
        "parameters": [
          {
            "name": "clinicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "example": "\"0\""
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "administrativeReference"
                ],
                "properties": {
                  "administrativeReference": {
                    "type": "string",
                    "nullable": true,
                    "minLength": 1,
                    "maxLength": 40
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicPatientAdministrativeReferenceDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed headers or strict DTO."
          },
          "401": {
            "description": "Clinic employee JWT is required."
          },
          "403": {
            "description": "Mutation capability is required."
          },
          "404": {
            "description": "Rollout disabled or patient unavailable in the exact scope."
          },
          "409": {
            "description": "Stale version, collision, association change or key reuse."
          },
          "422": {
            "description": "Invalid administrative reference."
          },
          "428": {
            "description": "A strong If-Match is required."
          },
          "503": {
            "description": "Visibility policy is unavailable."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Set or clear an exact-scope clinic patient administrative reference",
        "tags": [
          "Clinic Patients"
        ],
        "x-required-capabilities": [
          "patient.admin.local-profile.update"
        ]
      }
    },
...
  },
  "info": {
    "title": "VetHelp Booking Core API",
    "description": "Контракт локального удержания слотов, подтверждения записи клиникой и освобождения hold.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "opaque Owner session or legacy staff JWT",
        "type": "http",
        "description": "Opaque vh_ Owner session or compatibility JWT for staff/system roles. Authorization and expiry are evaluated server-side for Owner sessions."
      }
    },
    "schemas": {
...
      "ClinicPatientPetDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "speciesLabel": {
            "type": "string"
          },
          "breed": {
            "type": "string",
            "nullable": true
          },
          "sexCode": {
            "type": "string",
            "nullable": true,
            "enum": [
              "MALE",
              "FEMALE",
              "UNKNOWN"
            ]
          },
          "birthDate": {
            "type": "string",
            "nullable": true,
            "format": "date"
          }
        },
        "required": [
          "displayName",
          "speciesLabel",
          "breed",
          "sexCode",
          "birthDate"
        ]
      },
      "ClinicPatientOwnerDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "displayName"
        ]
      },
      "ClinicPatientRelationshipDto": {
        "type": "object",
        "properties": {
          "firstSeenAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastSeenAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "firstSeenAt",
          "lastSeenAt"
        ]
      },
...
      "ClinicPatientAppointmentSummaryServiceDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "displayName"
        ]
      },
      "ClinicPatientAppointmentSummaryVeterinarianDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "displayName"
        ]
      },
      "ClinicPatientAppointmentSummaryDto": {
        "type": "object",
        "properties": {
          "appointmentId": {
            "type": "string",
            "format": "uuid"
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "statusCode": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "COMPLETED",
              "NO_SHOW",
              "CANCELLED",
              "UNKNOWN"
            ]
          },
          "statusLabel": {
            "type": "string"
          },
          "service": {
            "$ref": "#/components/schemas/ClinicPatientAppointmentSummaryServiceDto"
          },
          "veterinarian": {
            "$ref": "#/components/schemas/ClinicPatientAppointmentSummaryVeterinarianDto"
          }
        },
        "required": [
          "appointmentId",
          "startsAt",
          "endsAt",
          "statusCode",
          "statusLabel",
          "service",
          "veterinarian"
        ]
      },
      "ClinicPatientDetailAppointmentsDto": {
        "type": "object",
        "properties": {
          "last": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ClinicPatientAppointmentSummaryDto"
              }
            ]
          },
          "next": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ClinicPatientAppointmentSummaryDto"
              }
            ]
          },
          "recent": {
            "maxItems": 10,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClinicPatientAppointmentSummaryDto"
            }
          }
        },
        "required": [
          "last",
          "next",
          "recent"
        ]
      },
      "ClinicPatientDetailPatientDto": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "pet": {
            "$ref": "#/components/schemas/ClinicPatientPetDto"
          },
          "owner": {
            "$ref": "#/components/schemas/ClinicPatientOwnerDto"
          },
          "relationship": {
            "$ref": "#/components/schemas/ClinicPatientRelationshipDto"
          },
          "appointments": {
            "$ref": "#/components/schemas/ClinicPatientDetailAppointmentsDto"
          },
          "localProfile": {
            "allOf": [
              {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "alias",
                  "administrativeReference",
                  "aggregateVersion",
                  "updatedAt"
                ],
                "properties": {
                  "alias": {
                    "type": "string",
                    "nullable": true
                  },
                  "administrativeReference": {
                    "type": "string",
                    "nullable": true
                  },
                  "aggregateVersion": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "updatedAt": {
                    "type": "string",
                    "format": "date-time",
                    "nullable": true
                  }
                }
              }
            ]
          }
        },
        "required": [
          "patientId",
          "pet",
          "owner",
          "relationship",
          "appointments",
          "localProfile"
        ]
      },
      "ClinicPatientDetailDto": {
        "type": "object",
        "properties": {
          "clinicId": {
            "type": "string",
            "format": "uuid"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "serverNow": {
            "type": "string",
            "format": "date-time"
          },
          "patient": {
            "$ref": "#/components/schemas/ClinicPatientDetailPatientDto"
          }
        },
        "required": [
          "clinicId",
          "locationId",
          "serverNow",
          "patient"
        ]
      },
...
      "ClinicPatientAdministrativeReferenceDto": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "clinicId": {
            "type": "string",
            "format": "uuid"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "alias": {
            "type": "string",
            "nullable": true
          },
          "administrativeReference": {
            "type": "string",
            "nullable": true
          },
          "aggregateVersion": {
            "type": "number",
            "minimum": 1
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "patientId",
          "clinicId",
          "locationId",
          "alias",
          "administrativeReference",
          "aggregateVersion",
          "updatedAt"
        ]
      },
...
    }
  }
}